StampedEntity_Timestamp_to_Long

+++ StateOfNeo/StateOfNeo.Data.Models/StampedEntity.cs
@@ -2,6 +2,6 @@
 {
     public abstract class StampedEntity : BaseEntity
     {
-        public uint Timestamp { get; set; }
+        public long Timestamp { get; set; }
     }
 }
